+++ tests/Feature/Controller/Student/StudentDetailControllerTest.php
@@ -1,13 +1,9 @@
 <?php
 
-declare(strict_types=1);
-
 namespace Tests\Feature\Controller\Student;
 
 use Illuminate\Foundation\Testing\DatabaseTransactions;
 use Tests\TestCase;
-use App\Http\Controllers\api\Student\StudentDetailController;
-use App\Service\Student\StudentDetailService;
 use App\Models\Student;
 
 class StudentDetailControllerTest extends TestCase
@@ -22,71 +18,39 @@ class StudentDetailControllerTest extends TestCase
         parent::setUp();
 
         $this->student = Student::factory()->create();
-
         $this->resume = $this->student->resume()->create();
     }
-    
+
     public function testStudentDetailsAreFound(): void
     {
-        $studentDetailService = $this->createMock(StudentDetailService::class);
-
-        $student = $this->student;
-    
-        $studentId = $student->id;
-        
-        $studentDetail = $this->resume;
-
-        $studentDetailService->expects($this->once())
-                              ->method('execute')
-                              ->with($studentId)
-                              ->willReturn($studentDetail->toArray());
-
-        $this->app->instance(StudentDetailService::class, $studentDetailService);
-
-        $response = $this->get(route('student.details', ['studentId' => $studentId]));
+        $response = $this->get(route('student.details', ['student' => $this->student]));
 
         $response->assertStatus(200);
 
-        $response->assertJsonStructure();
+        $response->assertJson([
+            'id' => $this->student->id,
+        ]);
     }
 
     public function testStudentDetailsIsNotFound(): void
     {
-        $studentDetailService = $this->createMock(StudentDetailService::class);
+        $nonExistentStudentId = 12345;
 
-        $studentId = 12345;
-
-        $studentDetailService->expects($this->once())
-                              ->method('execute')
-                              ->with($studentId)
-                              ->willThrowException(new \App\Exceptions\StudentNotFoundException((string)$studentId));
-
-        $this->app->instance(StudentDetailService::class, $studentDetailService);
-
-        $response = $this->get(route('student.details', ['studentId' => $studentId]));
+        $response = $this->get(route('student.details', ['student' => $nonExistentStudentId]));
 
         $response->assertStatus(404);
-        
-        $response->assertJson(['message' => 'No s\'ha trobat cap estudiant amb aquest ID: 12345']);
+
+        $response->assertJson(['message' => 'No query results for model [App\\Models\\Student] ' . $nonExistentStudentId]);
     }
 
-    public function testStudentDetailControllerReturns_404StatusAndResumeNotFoundExceptionMessageForValidStudentUuidWithoutResume(): void
+    public function testStudentDetailControllerReturns_404StatusAndResumeNotFoundExceptionMessageForValidStudentWithoutResume(): void
     {
         $this->student->resume->delete();
 
-        $response = $this->get(route('student.details', ['studentId' => $this->student]));
+        $response = $this->get(route('student.details', ['student' => $this->student]));
 
         $response->assertStatus(404);
-        
+
         $response->assertJson(['message' => 'No s\'ha trobat cap currículum per a l\'estudiant amb id: ' . $this->student->id]);
-    }
-
-    public function testStudentDetailControllerCanBeInstantiated():void
-    {
-        $studentDetailService = $this->createMock(StudentDetailService::class);
-        
-        $controller = new StudentDetailController($studentDetailService);
-
-        $this->assertInstanceOf(StudentDetailController::class, $controller);
     }
 }
